:recycle:(trainer): move validation logic into genome.ts
Move validateGenome, repairGenome, and helpers from validation.ts into
genome.ts where they are co-located with the types they validate.
validation.ts now re-exports from genome.ts for backward compatibility.

Closes #167
